+++ lightly_studio/src/lightly_studio/resolvers/datasets_resolver/__init__.py
@@ -1,9 +1,31 @@
 """Resolvers for database operations."""
 
+from lightly_studio.resolvers.datasets_resolver.create import create
+from lightly_studio.resolvers.datasets_resolver.delete import delete
+from lightly_studio.resolvers.datasets_resolver.export import (
+    export,
+    get_filtered_samples_count,
+)
+from lightly_studio.resolvers.datasets_resolver.get_all import get_all
+from lightly_studio.resolvers.datasets_resolver.get_by_id import get_by_id
+from lightly_studio.resolvers.datasets_resolver.get_by_name import get_by_name
 from lightly_studio.resolvers.datasets_resolver.get_dataset_details import (
     get_dataset_details,
 )
+from lightly_studio.resolvers.datasets_resolver.get_hierarchy import (
+    get_hierarchy,
+)
+from lightly_studio.resolvers.datasets_resolver.update import update
 
 __all__ = [
+    "create",
+    "delete",
+    "export",
+    "get_all",
+    "get_by_id",
+    "get_by_name",
     "get_dataset_details",
+    "get_filtered_samples_count",
+    "get_hierarchy",
+    "update",
 ]